refactor(hooks): extract tokenization helpers to _hook_utils
Three PreToolUse(Bash) hooks reuse the same safe_tokenize / strip_prefix /
iter_command_starts pipeline plus the constants they consume. The TODO at
hooks/block-gh-state-all.py:25 named the third hook as the trigger
condition for extraction. Issue #139 (memory-hint hook) hits that trigger,
so move the helpers to a shared module before adding the new consumer.

- New: hooks/_hook_utils.py with the helpers + 5 supporting constants
- block-gh-state-all.py: drop duplicate code, import from _hook_utils
- side-effect-scan.py: drop duplicate code, import from _hook_utils

No behavior change. Verified by existing regression suites:
- bash hooks/test-block-gh-state-all.sh -> Passed: 22 Failed: 0
- bash tests/test_side_effect_scan.sh -> PASS: 54 FAIL: 0
