+++ hooks/_hook_utils.py
@@ -1,0 +1,148 @@
+"""Shared tokenization helpers for praxis PreToolUse(Bash) hooks.
+
+Three hooks reuse the exact same `safe_tokenize` / `strip_prefix` /
+`iter_command_starts` pipeline plus the constants they consume:
+
+- `block-gh-state-all.py` — blocks invalid `gh search ... --state all`
+- `side-effect-scan.py` — surfaces collateral side effects via `ask`
+- `memory-hint.py` — emits stderr hints for `hookable: true` memories
+
+Per the TODO that lived at `block-gh-state-all.py:25`, this module is the
+extraction triggered by the third consumer (issue #139). The helpers move
+verbatim — see git history for the prior duplicated copies.
+"""
+from __future__ import annotations
+
+import re
+import shlex
+
+
+SHELL_SEPARATORS = {";", "&&", "||", "|", "&"}
+
+ENV_ASSIGN_RE = re.compile(r"^[A-Za-z_][A-Za-z0-9_]*=")
+
+# Shell keywords that appear at the start of a command segment but are purely
+# syntactic. `if true; then git push; fi` segments as `['if','true']`,
+# `['then','git','push']`, `['fi']` — we peel the keyword so argv[0] becomes
+# the real executable.
+SHELL_KEYWORDS = {
+    "if", "then", "elif", "else", "fi",
+    "while", "until", "do", "done",
+    "case", "esac", "in", "for",
+    "{", "}", "!", "function",
+}
+
+# Prefix wrappers that execute the following command as a new process. The
+# scanner looks past them to find the real argv[0]. Per-wrapper option
+# dictionaries list *only* flags that take a separate-token argument so that
+# `sudo --user admin kubectl ...` peels both `--user` and `admin`. Bare flags
+# (with no arg) and `--long=value` forms are handled generically below.
+PREFIX_WRAPPERS = {"env", "sudo", "nice", "time", "stdbuf", "ionice"}
+WRAPPER_OPTS_WITH_ARG = {
+    "env": {"-u", "--unset", "-C", "--chdir", "-S", "--split-string"},
+    "sudo": {
+        "-u", "-g", "-p", "-C", "-D", "-r", "-t", "-T", "-U", "-h",
+        "--user", "--group", "--prompt", "--close-from", "--chdir",
+        "--role", "--type", "--host", "--other-user",
+    },
+    "nice": {"-n", "--adjustment"},
+    "stdbuf": {"-i", "-o", "-e", "--input", "--output", "--error"},
+    "time": {"-f", "--format", "-o", "--output"},
+    "ionice": {
+        "-c", "--class", "-n", "--classdata",
+        "-p", "--pid", "-P", "--pgid", "-u", "--uid",
+    },
+}
+
+
+def safe_tokenize(command: str) -> list[str]:
+    """Tokenize with shell operators and line breaks split into tokens.
+
+    Uses shlex.shlex(punctuation_chars=';|&') so that `git push&&echo` and
+    `git push;echo` split into `['git', 'push', '&&', 'echo']` etc. Plain
+    shlex.split keeps operators glued to adjacent words, which would let a
+    whitespace-free one-liner bypass detection entirely.
+
+    Newlines are a command separator in Bash but shlex's whitespace_split
+    consumes them as generic whitespace, flattening multi-line scripts into
+    one token stream. We pre-split the raw command on `\\n` and insert a
+    synthetic `;` between line tokens so iter_command_starts sees the break.
+    Lines that fail to parse (unmatched quote, runaway heredoc, etc.) are
+    skipped — better a silent pass than a crashed hook.
+    """
+    lines = [ln for ln in command.split("\n") if ln.strip()]
+    if not lines:
+        return []
+    tokens: list[str] = []
+    for idx, line in enumerate(lines):
+        if idx > 0:
+            tokens.append(";")
+        try:
+            lex = shlex.shlex(line, posix=True, punctuation_chars=";|&")
+            lex.whitespace_split = True
+            lex.commenters = ""  # raw `#` is not a comment here; opt-out marker
+            tokens.extend(list(lex))
+        except ValueError:
+            continue
+    return tokens
+
+
+def strip_prefix(argv: list[str]) -> list[str]:
+    """Peel shell keywords, `KEY=VAL` assignments, and wrapper commands off
+    the front so argv[0] becomes the real executable.
+
+    Handles (in any order, iteratively):
+    - shell keywords (`if`, `then`, `do`, `while`, etc.) — pure syntax, drop
+    - env assignments (`FOO=1`) — drop
+    - wrapper commands (`env`, `sudo`, `nice`, `time`, `stdbuf`, `ionice`) —
+      drop the wrapper plus its option flags. Option flags are peeled
+      generically: any `-*` token is consumed, and if it's a known arg-taking
+      flag for this wrapper the following value token is peeled too. The
+      `--long=value` form counts as a single token and is handled naturally.
+    """
+    i = 0
+    n = len(argv)
+    while i < n:
+        tok = argv[i]
+        if tok in SHELL_KEYWORDS:
+            i += 1
+            continue
+        if ENV_ASSIGN_RE.match(tok):
+            i += 1
+            continue
+        if tok in PREFIX_WRAPPERS:
+            wrapper = tok
+            i += 1
+            opts_with_arg = WRAPPER_OPTS_WITH_ARG.get(wrapper, set())
+            while i < n:
+                nxt = argv[i]
+                if ENV_ASSIGN_RE.match(nxt):
+                    i += 1
+                    continue
+                if not nxt.startswith("-"):
+                    break
+                if "=" in nxt:
+                    # --long=value — value embedded; peel this token only
+                    i += 1
+                    continue
+                if nxt in opts_with_arg and i + 1 < n:
+                    # --user admin / -u admin — peel pair
+                    i += 2
+                    continue
+                # bare flag (-E, -i, -oL, etc.) — peel single token
+                i += 1
+            continue
+        break
+    return argv[i:]
+
+
+def iter_command_starts(tokens: list[str]):
+    """Yield argv slices at each command start across shell separators."""
+    start = 0
+    for i, tok in enumerate(tokens):
+        if tok in SHELL_SEPARATORS:
+            if start < i:
+                yield tokens[start:i]
+            start = i + 1
+    if start < len(tokens):
+        yield tokens[start:]
